feat(admin): add cache monitoring tab
