+++ apps/admin/src/pages/Monitoring.tsx
@@ -2,14 +2,14 @@ import { useState } from "react";
 
 import RumTab from "../features/monitoring/RumTab";
 import RateLimitsTab from "../features/monitoring/RateLimitsTab";
-import CacheTools from "./CacheTools";
+import CacheTab from "../features/monitoring/CacheTab";
 import AuditLog from "./AuditLog";
 import Jobs from "./Jobs";
 
 const tabs = [
   { id: "rum", label: "RUM", component: <RumTab /> },
   { id: "rate-limits", label: "Rate limits", component: <RateLimitsTab /> },
-  { id: "cache", label: "Cache", component: <CacheTools /> },
+  { id: "cache", label: "Cache", component: <CacheTab /> },
   { id: "audit-log", label: "Audit log", component: <AuditLog /> },
   { id: "jobs", label: "Jobs", component: <Jobs /> },
 ] as const;
